move ShoppingItem from interfaces to types
