+++ smart-cart/src/types/ShoppingItem.tsx
